refactor: 패키지 이동
- 출석과 관련된 정책, 결석과 관련된 정책으로 구분
