+++ src/main/java/domain/policy/attend/time/AttendanceTimePolicy.java
@@ -1,8 +1,6 @@
-package domain.policy.time;
+package domain.policy.attend.time;
 
-import domain.policy.time.rule.AttendanceStateRule;
-import domain.policy.time.rule.CampusTimeRule;
-import domain.policy.time.rule.StudyTimeRule;
+import domain.policy.AttendanceStateRule;
 
 import java.time.LocalTime;
