+++ src/main/java/domain/policy/attend/time/CampusTimeRule.java
@@ -1,4 +1,4 @@
-package domain.policy.time.rule;
+package domain.policy.attend.time;
 
 import java.time.LocalTime;
