+++ src/test/java/domain/AttendanceBookTest.java
@@ -1,8 +1,8 @@
 package domain;
 
-import domain.policy.AttendancePolicy;
-import domain.policy.date.AttendanceDatePolicy;
-import domain.policy.time.AttendanceTimePolicy;
+import domain.policy.attend.AttendancePolicy;
+import domain.policy.attend.date.AttendanceDatePolicy;
+import domain.policy.attend.time.AttendanceTimePolicy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import reader.AttendanceFileReader;
